backend/requirements.txt
fastapi==0.95.2
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
slowapi==0.1.7
limits==2.3.1
redis==4.5.5